requirements.txt
duneanalytics==1.2.0
matplotlib==3.5.2
python-dotenv==0.16.0
pandas==1.4.1
numpy==1.22.4
scipy==1.8.1